trabalho com validaçao de chave e cache de chave de acesso iniciado
